Fixed hp change after attack

+++ unity-client/Assets/Scenes/Game/Card/CardManager.cs
@@ -125,7 +125,7 @@ public class CardManager : MonoBehaviour
             cardDisplay.Untap();
         }
 
-        cardDisplay.CurrentHp = card.newHp;
+        cardDisplay.CurrentHp = (int)card.newHp;
 
         if (card.killed)
         {
